+++ assets/packages/core/tailwindcss-v3/play/observer.ts
@@ -19,7 +19,11 @@ let classes = new Set<string>()
 /**
  * The stylesheet that we use to inject the compiled CSS into the page.
  */
-let sheet = document.createElement('style')
+let sheet = document.querySelector('style#windpress-cached-inline-css') as HTMLStyleElement | null;
+if (!sheet) {
+    sheet = document.createElement('style');
+    sheet.id = 'windpress-cached-inline-css';
+}
 
 /**
  * The queue of build tasks that need to be run. This is used to ensure that we
@@ -100,9 +104,9 @@ async function build(kind: 'full' | 'incremental') {
     // if (newClasses.size === 0 && kind === 'incremental') return
 
     // 2. Compile the CSS
-    I.start(`Build utilities`)
+    I.start(`Build utilities`);
 
-    sheet.textContent = await _compile({
+    (sheet as HTMLStyleElement).textContent = await _compile({
         entrypoint: {
             css: '/main.css',
             config: '/tailwind.config.js',